Add progress bar to pandas apply
